display-controller/Makefile: Improvements an PLL config generation
Improve a number of features, including the generation of log files for
synthesis, place and route, timing as well as for the new PLL config
generation. Some of this logs are automatically reported when building
with V=1 otherwise they must be accessed in the obj/*.log files.

The timing analysis now reports pass/fail.

Specific defines are now added for the device and target frequencies.